# Eft_grupo14/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MiniMarket Plus API
  version: 1.0.0
  description: API EFT para autenticacion, inventario, pedidos, promociones y reportes.
servers:
  - url: http://localhost:8080
security:
  - bearer-jwt: []
components:
  securitySchemes:
    bearer-jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/inventory/products:
    post:
      summary: Crea producto de inventario.
      responses:
        "201":
          description: Producto creado.
        "403":
          description: Requiere rol ADMIN.
  /api/inventory/products/{id}/stock:
    patch:
      summary: Actualiza stock y genera orden de compra si llega al minimo.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stock actualizado.
...